Extract Different Rows - User-facing Function to extract missing or new rows from comparing two dataframes.

Extract Different Rows - User-facing Function to extract missing or new rows from comparing two dataframes.

Usage

extract_diff_rows(
  dfx = NULL,
  dfy = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  myrror_object = NULL,
  output = c("simple", "full", "silent")
)

Arguments

dfx

data.frame object.

dfy

data.frame object.

by

character, key to be used for dfx and dfy.

by.x

character, key to be used for dfx.

by.y

character, key to be used for dfy.

myrror_object

myrror object.

output

character, one of "simple", "full", "silent".

Value

data.table object with the rows that are missing or new.

Examples

extract_diff_rows(iris, iris_var1)
#>        rn Sepal.Length.x Sepal.Width.x Petal.Length.x Petal.Width.x Species.x
#>    <char>          <num>         <num>          <num>         <num>    <char>
#> 1:    151             NA            NA             NA            NA      <NA>
#> 2:    152             NA            NA             NA            NA      <NA>
#> 3:    153             NA            NA             NA            NA      <NA>
#> 4:    154             NA            NA             NA            NA      <NA>
#> 5:    155             NA            NA             NA            NA      <NA>
#>    Sepal.Length.y Sepal.Width.y Petal.Length.y Petal.Width.y Species.y
#>             <num>         <num>          <num>         <num>    <char>
#> 1:            5.1           3.5            1.4           0.2    setosa
#> 2:            4.9           3.0            1.4           0.2    setosa
#> 3:            4.7           3.2            1.3           0.2    setosa
#> 4:            4.6           3.1            1.5           0.2    setosa
#> 5:            5.0           3.6            1.4           0.2    setosa
#>    Petal.Area  .joyn row_index
#>         <num> <fctr>     <int>
#> 1:       0.28      y        NA
#> 2:       0.28      y        NA
#> 3:       0.26      y        NA
#> 4:       0.30      y        NA
#> 5:       0.28      y        NA
extract_diff_rows(survey_data, survey_data_2, by=c('country', 'year'))
#> Empty data.table (0 rows and 13 cols): country,year,variable1.x,variable2.x,variable3.x,variable4.x...