Extract Different Values - User-facing - Table format Function to extract rows with different values between two dataframes.

Extract Different Values - User-facing - Table format Function to extract rows with different values between two dataframes.

Usage

extract_diff_table(
  dfx = NULL,
  dfy = NULL,
  myrror_object = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  output = c("simple", "full", "silent"),
  tolerance = 1e-07,
  verbose = TRUE
)

Arguments

dfx

a non-empty data.frame.

dfy

a non-empty data.frame.

myrror_object

myrror object from create_myrror_object

by

character, key to be used for dfx and dfy.

by.x

character, key to be used for dfx.

by.y

character, key to be used for dfy.

output

character: one of "full", "simple", "silent".

tolerance

numeric, default to 1e-7.

verbose

logical: If TRUE additional information will be displayed.

Value

data.table object with all observations for which at least 1 value is different.

Examples

# 1. Standard report, after running myrror() or compare_values():
myrror(survey_data, survey_data_2, by=c('country', 'year'))
#> 
#> ── Myrror Report ───────────────────────────────────────────────────────────────
#> 
#> ── General Information: ──
#> 
#> dfx: survey_data with 16 rows and 6 columns.
#> dfy: survey_data_2 with 16 rows and 6 columns.
#> Keys: country and year.
#> 
#> ── Note: comparison is done for shared columns and rows. ──
#> 
#> ✔ Total shared columns (no keys): 4
#> ! Non-shared columns in survey_data: 2 (country and year)
#> ! Non-shared columns in survey_data_2: 2 ()
#> 
#> ✔ Total shared rows: 16
#> ! Non-shared rows in survey_data: 0.
#> ! Non-shared rows in survey_data_2: 0.
#> 
#> ── 1. Shared Columns Class Comparison ──────────────────────────────────────────
#> 
#> ✔ All shared columns have the same class.
#> 
#> 
#> ── 2. Shared Columns Values Comparison ─────────────────────────────────────────
#> 
#> ! 1 shared column(s) have different value(s):
#> ℹ Note: character-numeric comparison is allowed.
#> 
#> 
#> ── Overview: ──
#> 
#> # A tibble: 1 × 4
#>   variable  change_in_value na_to_value value_to_na
#>   <fct>               <int>       <int>       <int>
#> 1 variable2              16           0           0
#> 
#> 
#> 
#> ── Value comparison: ──
#> 
#> ! 1 shared column(s) have different value(s):
#> 
#> ── variable2 
#>                diff indexes variable2.x variable2.y
#>              <char>  <char>       <num>       <num>
#>  1: change_in_value       1  -1.0974903   2.2530832
#>  2: change_in_value      10  -0.2509333   0.7871750
#>  3: change_in_value      11  -1.9006292   1.1134531
#>  4: change_in_value      12  -0.7021324   1.0855454
#>  5: change_in_value      13   0.6587007  -0.4379345
#>  6: change_in_value      14   3.1408877  -0.4860943
#>  7: change_in_value      15   0.1652734   0.7041218
#>  8: change_in_value      16  -0.9723386  -1.4740967
#>  9: change_in_value       2  -0.4876537   1.8377825
#> 10: change_in_value       3  -0.1286307  -0.2922308
#> 11: change_in_value       4  -0.5524387  -0.5077830
#> 12: change_in_value       5  -1.1208384   0.2812097
#> 13: change_in_value       6  -0.8406131   0.5614637
#> 14: change_in_value       7  -0.7293604   0.9611788
#> 15: change_in_value       8  -1.8557837  -0.6865351
#> 16: change_in_value       9   0.8019353  -0.5732589
#> 
#> ℹ Note: run `extract_diff_values()` or `extract_diff_table()` to access the results in list or table format.
#> 
#> ✔ End of Myrror Report.
extract_diff_table()
#> Last myrror object used for comparison.
#>                diff indexes country  year variable1.x variable2.x  variable3.x
#>              <char>  <char>  <char> <int>       <num>       <num>        <num>
#>  1: change_in_value       1       A  2010  0.92567152  -1.0974903 -0.330966991
#>  2: change_in_value       2       A  2011 -0.11439247  -0.4876537 -0.675403198
#>  3: change_in_value       3       A  2012 -0.02104358  -0.1286307 -1.500701944
#>  4: change_in_value       4       A  2013 -0.06520351  -0.5524387 -1.330601404
#>  5: change_in_value       5       A  2014 -0.72294750  -1.1208384 -0.200692824
#>  6: change_in_value       6       A  2015 -0.30728314  -0.8406131  0.002919924
#>  7: change_in_value       7       A  2016  1.69292174  -0.7293604  0.526351858
#>  8: change_in_value       8       A  2017 -0.88575936  -1.8557837 -0.436418266
#>  9: change_in_value       9       B  2010  0.14322693   0.8019353  0.357916400
#> 10: change_in_value      10       B  2011 -0.31830385  -0.2509333  1.445748740
#> 11: change_in_value      11       B  2012 -1.07735583  -1.9006292 -0.129643815
#> 12: change_in_value      12       B  2013  1.22851017  -0.7021324 -0.603758741
#> 13: change_in_value      13       B  2014 -0.88046550   0.6587007 -0.595608958
#> 14: change_in_value      14       B  2015  0.85375405   3.1408877 -1.073222461
#> 15: change_in_value      15       B  2016  1.32535591   0.1652734  0.718469591
#> 16: change_in_value      16       B  2017  0.42100237  -0.9723386  0.447581638
#>     variable4.x variable1.y variable2.y  variable3.y variable4.y  .joyn
#>           <num>       <num>       <num>        <num>       <num> <fctr>
#>  1: -1.18260779  0.92567152   2.2530832 -0.330966991 -1.18260779  x & y
#>  2:  0.06448929 -0.11439247   1.8377825 -0.675403198  0.06448929  x & y
#>  3:  0.08787883 -0.02104358  -0.2922308 -1.500701944  0.08787883  x & y
#>  4: -0.28839409 -0.06520351  -0.5077830 -1.330601404 -0.28839409  x & y
#>  5: -1.21014163 -0.72294750   0.2812097 -0.200692824 -1.21014163  x & y
#>  6:  0.62950207 -0.30728314   0.5614637  0.002919924  0.62950207  x & y
#>  7:  2.01149703  1.69292174   0.9611788  0.526351858  2.01149703  x & y
#>  8:  0.38265713 -0.88575936  -0.6865351 -0.436418266  0.38265713  x & y
#>  9: -0.97376153  0.14322693  -0.5732589  0.357916400 -0.97376153  x & y
#> 10:  0.45145400 -0.31830385   0.7871750  1.445748740  0.45145400  x & y
#> 11: -0.28450993 -1.07735583   1.1134531 -0.129643815 -0.28450993  x & y
#> 12:  0.76454938  1.22851017   1.0855454 -0.603758741  0.76454938  x & y
#> 13: -0.92083472 -0.88046550  -0.4379345 -0.595608958 -0.92083472  x & y
#> 14:  0.55352418  0.85375405  -0.4860943 -1.073222461  0.55352418  x & y
#> 15:  0.21744978  1.32535591   0.7041218  0.718469591  0.21744978  x & y
#> 16: -0.36628320  0.42100237  -1.4740967  0.447581638 -0.36628320  x & y
#>     row_index
#>         <int>
#>  1:         1
#>  2:         2
#>  3:         3
#>  4:         4
#>  5:         5
#>  6:         6
#>  7:         7
#>  8:         8
#>  9:         9
#> 10:        10
#> 11:        11
#> 12:        12
#> 13:        13
#> 14:        14
#> 15:        15
#> 16:        16

# 2. Standard report, with new data:
extract_diff_table(survey_data, survey_data_2, by=c('country', 'year'))
#>                diff indexes country  year variable1.x variable2.x  variable3.x
#>              <char>  <char>  <char> <int>       <num>       <num>        <num>
#>  1: change_in_value       1       A  2010  0.92567152  -1.0974903 -0.330966991
#>  2: change_in_value       2       A  2011 -0.11439247  -0.4876537 -0.675403198
#>  3: change_in_value       3       A  2012 -0.02104358  -0.1286307 -1.500701944
#>  4: change_in_value       4       A  2013 -0.06520351  -0.5524387 -1.330601404
#>  5: change_in_value       5       A  2014 -0.72294750  -1.1208384 -0.200692824
#>  6: change_in_value       6       A  2015 -0.30728314  -0.8406131  0.002919924
#>  7: change_in_value       7       A  2016  1.69292174  -0.7293604  0.526351858
#>  8: change_in_value       8       A  2017 -0.88575936  -1.8557837 -0.436418266
#>  9: change_in_value       9       B  2010  0.14322693   0.8019353  0.357916400
#> 10: change_in_value      10       B  2011 -0.31830385  -0.2509333  1.445748740
#> 11: change_in_value      11       B  2012 -1.07735583  -1.9006292 -0.129643815
#> 12: change_in_value      12       B  2013  1.22851017  -0.7021324 -0.603758741
#> 13: change_in_value      13       B  2014 -0.88046550   0.6587007 -0.595608958
#> 14: change_in_value      14       B  2015  0.85375405   3.1408877 -1.073222461
#> 15: change_in_value      15       B  2016  1.32535591   0.1652734  0.718469591
#> 16: change_in_value      16       B  2017  0.42100237  -0.9723386  0.447581638
#>     variable4.x variable1.y variable2.y  variable3.y variable4.y  .joyn
#>           <num>       <num>       <num>        <num>       <num> <fctr>
#>  1: -1.18260779  0.92567152   2.2530832 -0.330966991 -1.18260779  x & y
#>  2:  0.06448929 -0.11439247   1.8377825 -0.675403198  0.06448929  x & y
#>  3:  0.08787883 -0.02104358  -0.2922308 -1.500701944  0.08787883  x & y
#>  4: -0.28839409 -0.06520351  -0.5077830 -1.330601404 -0.28839409  x & y
#>  5: -1.21014163 -0.72294750   0.2812097 -0.200692824 -1.21014163  x & y
#>  6:  0.62950207 -0.30728314   0.5614637  0.002919924  0.62950207  x & y
#>  7:  2.01149703  1.69292174   0.9611788  0.526351858  2.01149703  x & y
#>  8:  0.38265713 -0.88575936  -0.6865351 -0.436418266  0.38265713  x & y
#>  9: -0.97376153  0.14322693  -0.5732589  0.357916400 -0.97376153  x & y
#> 10:  0.45145400 -0.31830385   0.7871750  1.445748740  0.45145400  x & y
#> 11: -0.28450993 -1.07735583   1.1134531 -0.129643815 -0.28450993  x & y
#> 12:  0.76454938  1.22851017   1.0855454 -0.603758741  0.76454938  x & y
#> 13: -0.92083472 -0.88046550  -0.4379345 -0.595608958 -0.92083472  x & y
#> 14:  0.55352418  0.85375405  -0.4860943 -1.073222461  0.55352418  x & y
#> 15:  0.21744978  1.32535591   0.7041218  0.718469591  0.21744978  x & y
#> 16: -0.36628320  0.42100237  -1.4740967  0.447581638 -0.36628320  x & y
#>     row_index
#>         <int>
#>  1:         1
#>  2:         2
#>  3:         3
#>  4:         4
#>  5:         5
#>  6:         6
#>  7:         7
#>  8:         8
#>  9:         9
#> 10:        10
#> 11:        11
#> 12:        12
#> 13:        13
#> 14:        14
#> 15:        15
#> 16:        16

# 3. Toggle tolerance:
extract_diff_table(survey_data, survey_data_2, by=c('country', 'year'),
                    tolerance = 1e-5)
#>                diff indexes country  year variable1.x variable2.x  variable3.x
#>              <char>  <char>  <char> <int>       <num>       <num>        <num>
#>  1: change_in_value       1       A  2010  0.92567152  -1.0974903 -0.330966991
#>  2: change_in_value       2       A  2011 -0.11439247  -0.4876537 -0.675403198
#>  3: change_in_value       3       A  2012 -0.02104358  -0.1286307 -1.500701944
#>  4: change_in_value       4       A  2013 -0.06520351  -0.5524387 -1.330601404
#>  5: change_in_value       5       A  2014 -0.72294750  -1.1208384 -0.200692824
#>  6: change_in_value       6       A  2015 -0.30728314  -0.8406131  0.002919924
#>  7: change_in_value       7       A  2016  1.69292174  -0.7293604  0.526351858
#>  8: change_in_value       8       A  2017 -0.88575936  -1.8557837 -0.436418266
#>  9: change_in_value       9       B  2010  0.14322693   0.8019353  0.357916400
#> 10: change_in_value      10       B  2011 -0.31830385  -0.2509333  1.445748740
#> 11: change_in_value      11       B  2012 -1.07735583  -1.9006292 -0.129643815
#> 12: change_in_value      12       B  2013  1.22851017  -0.7021324 -0.603758741
#> 13: change_in_value      13       B  2014 -0.88046550   0.6587007 -0.595608958
#> 14: change_in_value      14       B  2015  0.85375405   3.1408877 -1.073222461
#> 15: change_in_value      15       B  2016  1.32535591   0.1652734  0.718469591
#> 16: change_in_value      16       B  2017  0.42100237  -0.9723386  0.447581638
#>     variable4.x variable1.y variable2.y  variable3.y variable4.y  .joyn
#>           <num>       <num>       <num>        <num>       <num> <fctr>
#>  1: -1.18260779  0.92567152   2.2530832 -0.330966991 -1.18260779  x & y
#>  2:  0.06448929 -0.11439247   1.8377825 -0.675403198  0.06448929  x & y
#>  3:  0.08787883 -0.02104358  -0.2922308 -1.500701944  0.08787883  x & y
#>  4: -0.28839409 -0.06520351  -0.5077830 -1.330601404 -0.28839409  x & y
#>  5: -1.21014163 -0.72294750   0.2812097 -0.200692824 -1.21014163  x & y
#>  6:  0.62950207 -0.30728314   0.5614637  0.002919924  0.62950207  x & y
#>  7:  2.01149703  1.69292174   0.9611788  0.526351858  2.01149703  x & y
#>  8:  0.38265713 -0.88575936  -0.6865351 -0.436418266  0.38265713  x & y
#>  9: -0.97376153  0.14322693  -0.5732589  0.357916400 -0.97376153  x & y
#> 10:  0.45145400 -0.31830385   0.7871750  1.445748740  0.45145400  x & y
#> 11: -0.28450993 -1.07735583   1.1134531 -0.129643815 -0.28450993  x & y
#> 12:  0.76454938  1.22851017   1.0855454 -0.603758741  0.76454938  x & y
#> 13: -0.92083472 -0.88046550  -0.4379345 -0.595608958 -0.92083472  x & y
#> 14:  0.55352418  0.85375405  -0.4860943 -1.073222461  0.55352418  x & y
#> 15:  0.21744978  1.32535591   0.7041218  0.718469591  0.21744978  x & y
#> 16: -0.36628320  0.42100237  -1.4740967  0.447581638 -0.36628320  x & y
#>     row_index
#>         <int>
#>  1:         1
#>  2:         2
#>  3:         3
#>  4:         4
#>  5:         5
#>  6:         6
#>  7:         7
#>  8:         8
#>  9:         9
#> 10:        10
#> 11:        11
#> 12:        12
#> 13:        13
#> 14:        14
#> 15:        15
#> 16:        16