Extract Different Values - User-facing - List format Function to extract rows with different values between two dataframes.

Extract Different Values - User-facing - List format Function to extract rows with different values between two dataframes.

Usage

extract_diff_values(
  dfx = NULL,
  dfy = NULL,
  myrror_object = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  output = c("simple", "full", "silent"),
  tolerance = 1e-07,
  verbose = TRUE
)

Arguments

dfx

a non-empty data.frame.

dfy

a non-empty data.frame.

myrror_object

myrror object from create_myrror_object

by

character, key to be used for dfx and dfy.

by.x

character, key to be used for dfx.

by.y

character, key to be used for dfy.

output

character: one of "full", "simple", "silent"

tolerance

numeric, default to 1e-7.

verbose

logical: If TRUE additional information will be displayed

Value

list object with two items: diff_list and diff_table

Examples

extract_diff_values(iris, iris_var1)
#> $Sepal.Length
#>           diff indexes Sepal.Length.x Sepal.Length.y
#>         <char>  <char>          <num>          <num>
#> 1: value_to_na     104            6.3             NA
#> 2: value_to_na     125            6.7             NA
#> 3: value_to_na      67            5.6             NA
#> 4: value_to_na      80            5.7             NA
#> 5: value_to_na      96            5.7             NA
#> 
extract_diff_values(survey_data, survey_data_2, by=c('country', 'year'))
#> $variable2
#>                diff indexes variable2.x variable2.y
#>              <char>  <char>       <num>       <num>
#>  1: change_in_value       1 -0.36741872  0.48571921
#>  2: change_in_value      10 -0.08331262 -0.56094771
#>  3: change_in_value      11 -0.83435863 -0.62187200
#>  4: change_in_value      12  0.14181480 -0.45073978
#>  5: change_in_value      13 -1.27564125  0.02804128
#>  6: change_in_value      14 -0.06316779  0.92758765
#>  7: change_in_value      15 -1.28396057 -0.19687494
#>  8: change_in_value      16  0.73642976 -0.16630616
#>  9: change_in_value       2  1.08281824 -0.26755620
#> 10: change_in_value       3 -0.88608285 -1.22263500
#> 11: change_in_value       4 -1.35789652 -0.04244568
#> 12: change_in_value       5  0.49807310 -1.44511812
#> 13: change_in_value       6 -0.44915960 -0.87467620
#> 14: change_in_value       7 -1.35026961 -0.63024083
#> 15: change_in_value       8  0.10604084  1.05829679
#> 16: change_in_value       9  2.44402606 -0.77469057
#>