Extract Different Values - User-facing - Table format Function to extract rows with different values between two dataframes.

Extract Different Values - User-facing - Table format Function to extract rows with different values between two dataframes.

Usage

extract_diff_table(
  dfx = NULL,
  dfy = NULL,
  myrror_object = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  output = c("simple", "full", "silent"),
  tolerance = 1e-07,
  verbose = TRUE
)

Arguments

dfx

a non-empty data.frame.

dfy

a non-empty data.frame.

myrror_object

myrror object from create_myrror_object

by

character, key to be used for dfx and dfy.

by.x

character, key to be used for dfx.

by.y

character, key to be used for dfy.

output

character: one of "full", "simple", "silent".

tolerance

numeric, default to 1e-7.

verbose

logical: If TRUE additional information will be displayed.

Value

data.table object with all observations for which at least 1 value is different.

Examples

# 1. Standard report, after running myrror() or compare_values():
myrror(survey_data, survey_data_2, by=c('country', 'year'))
#> [1] 16
#> 
#> ── Myrror Report ───────────────────────────────────────────────────────────────
#> 
#> ── General Information: ──
#> 
#> dfx: survey_data with 16 rows and 6 columns.
#> dfy: survey_data_2 with 16 rows and 6 columns.
#> Keys: country and year.
#> 
#> ── Note: comparison is done for shared columns and rows. ──
#> 
#> ✔ Total shared columns (no keys): 4
#> ! Non-shared columns in survey_data: 2 (country and year)
#> ! Non-shared columns in survey_data_2: 2 ()
#> 
#> ✔ Total shared rows: 16
#> ! Non-shared rows in survey_data: 0.
#> ! Non-shared rows in survey_data_2: 0.
#> 
#> ── 1. Shared Columns Class Comparison ──────────────────────────────────────────
#> 
#> ✔ All shared columns have the same class.
#> 
#> 
#> ── 2. Shared Columns Values Comparison ─────────────────────────────────────────
#> 
#> ! 1 shared column(s) have different value(s):
#> ℹ Note: character-numeric comparison is allowed.
#> 
#> 
#> ── Overview: ──
#> 
#> # A tibble: 1 × 4
#>   variable  change_in_value na_to_value value_to_na
#>   <fct>               <int>       <int>       <int>
#> 1 variable2              16           0           0
#> 
#> 
#> 
#> ── Value comparison: ──
#> 
#> ! 1 shared column(s) have different value(s):
#> ℹ Note: Only first 5 rows shown for each variable.
#> 
#> ── "variable2" 
#>               diff indexes variable2.x variable2.y
#>             <char>  <char>       <num>       <num>
#> 1: change_in_value       1   0.4978505  -1.0717912
#> 2: change_in_value      10  -1.6866933  -0.7092008
#> 3: change_in_value      11   0.8377870  -0.6880086
#> 4: change_in_value      12   0.1533731   1.0255714
#> 5: change_in_value      13  -1.1381369  -0.2847730
#> ...
#> 
#> ℹ Note: run `extract_diff_values()` or `extract_diff_table()` to access the results in list or table format.
#> 
#> ✔ End of Myrror Report.
extract_diff_table()
#> Last myrror object used for comparison.
#> [1] 16
#>                diff indexes country  year variable1.x variable2.x variable3.x
#>              <char>  <char>  <char> <int>       <num>       <num>       <num>
#>  1: change_in_value       1       A  2010 -0.56047565   0.4978505  0.89512566
#>  2: change_in_value       2       A  2011 -0.23017749  -1.9666172  0.87813349
#>  3: change_in_value       3       A  2012  1.55870831   0.7013559  0.82158108
#>  4: change_in_value       4       A  2013  0.07050839  -0.4727914  0.68864025
#>  5: change_in_value       5       A  2014  0.12928774  -1.0678237  0.55391765
#>  6: change_in_value       6       A  2015  1.71506499  -0.2179749 -0.06191171
#>  7: change_in_value       7       A  2016  0.46091621  -1.0260044 -0.30596266
#>  8: change_in_value       8       A  2017 -1.26506123  -0.7288912 -0.38047100
#>  9: change_in_value       9       B  2010 -0.68685285  -0.6250393 -0.69470698
#> 10: change_in_value      10       B  2011 -0.44566197  -1.6866933 -0.20791728
#> 11: change_in_value      11       B  2012  1.22408180   0.8377870 -1.26539635
#> 12: change_in_value      12       B  2013  0.35981383   0.1533731  2.16895597
#> 13: change_in_value      13       B  2014  0.40077145  -1.1381369  1.20796200
#> 14: change_in_value      14       B  2015  0.11068272   1.2538149 -1.12310858
#> 15: change_in_value      15       B  2016 -0.55584113   0.4264642 -0.40288484
#> 16: change_in_value      16       B  2017  1.78691314  -0.2950715 -0.46665535
#>     variable4.x variable1.y variable2.y variable3.y variable4.y  .joyn
#>           <num>       <num>       <num>       <num>       <num> <fctr>
#>  1:  0.77996512 -0.56047565 -1.07179123  0.89512566  0.77996512  x & y
#>  2: -0.08336907 -0.23017749  0.30352864  0.87813349 -0.08336907  x & y
#>  3:  0.25331851  1.55870831  0.44820978  0.82158108  0.25331851  x & y
#>  4: -0.02854676  0.07050839  0.05300423  0.68864025 -0.02854676  x & y
#>  5: -0.04287046  0.12928774  0.92226747  0.55391765 -0.04287046  x & y
#>  6:  1.36860228  1.71506499  2.05008469 -0.06191171  1.36860228  x & y
#>  7: -0.22577099  0.46091621 -0.49103117 -0.30596266 -0.22577099  x & y
#>  8:  1.51647060 -1.26506123 -2.30916888 -0.38047100  1.51647060  x & y
#>  9: -1.54875280 -0.68685285  1.00573852 -0.69470698 -1.54875280  x & y
#> 10:  0.58461375 -0.44566197 -0.70920076 -0.20791728  0.58461375  x & y
#> 11:  0.12385424  1.22408180 -0.68800862 -1.26539635  0.12385424  x & y
#> 12:  0.21594157  0.35981383  1.02557137  2.16895597  0.21594157  x & y
#> 13:  0.37963948  0.40077145 -0.28477301  1.20796200  0.37963948  x & y
#> 14: -0.50232345  0.11068272 -1.22071771 -1.12310858 -0.50232345  x & y
#> 15: -0.33320738 -0.55584113  0.18130348 -0.40288484 -0.33320738  x & y
#> 16: -1.01857538  1.78691314 -0.13889136 -0.46665535 -1.01857538  x & y
#>     row_index
#>         <int>
#>  1:         1
#>  2:         2
#>  3:         3
#>  4:         4
#>  5:         5
#>  6:         6
#>  7:         7
#>  8:         8
#>  9:         9
#> 10:        10
#> 11:        11
#> 12:        12
#> 13:        13
#> 14:        14
#> 15:        15
#> 16:        16

# 2. Standard report, with new data:
extract_diff_table(survey_data, survey_data_2, by=c('country', 'year'))
#> [1] 16
#>                diff indexes country  year variable1.x variable2.x variable3.x
#>              <char>  <char>  <char> <int>       <num>       <num>       <num>
#>  1: change_in_value       1       A  2010 -0.56047565   0.4978505  0.89512566
#>  2: change_in_value       2       A  2011 -0.23017749  -1.9666172  0.87813349
#>  3: change_in_value       3       A  2012  1.55870831   0.7013559  0.82158108
#>  4: change_in_value       4       A  2013  0.07050839  -0.4727914  0.68864025
#>  5: change_in_value       5       A  2014  0.12928774  -1.0678237  0.55391765
#>  6: change_in_value       6       A  2015  1.71506499  -0.2179749 -0.06191171
#>  7: change_in_value       7       A  2016  0.46091621  -1.0260044 -0.30596266
#>  8: change_in_value       8       A  2017 -1.26506123  -0.7288912 -0.38047100
#>  9: change_in_value       9       B  2010 -0.68685285  -0.6250393 -0.69470698
#> 10: change_in_value      10       B  2011 -0.44566197  -1.6866933 -0.20791728
#> 11: change_in_value      11       B  2012  1.22408180   0.8377870 -1.26539635
#> 12: change_in_value      12       B  2013  0.35981383   0.1533731  2.16895597
#> 13: change_in_value      13       B  2014  0.40077145  -1.1381369  1.20796200
#> 14: change_in_value      14       B  2015  0.11068272   1.2538149 -1.12310858
#> 15: change_in_value      15       B  2016 -0.55584113   0.4264642 -0.40288484
#> 16: change_in_value      16       B  2017  1.78691314  -0.2950715 -0.46665535
#>     variable4.x variable1.y variable2.y variable3.y variable4.y  .joyn
#>           <num>       <num>       <num>       <num>       <num> <fctr>
#>  1:  0.77996512 -0.56047565 -1.07179123  0.89512566  0.77996512  x & y
#>  2: -0.08336907 -0.23017749  0.30352864  0.87813349 -0.08336907  x & y
#>  3:  0.25331851  1.55870831  0.44820978  0.82158108  0.25331851  x & y
#>  4: -0.02854676  0.07050839  0.05300423  0.68864025 -0.02854676  x & y
#>  5: -0.04287046  0.12928774  0.92226747  0.55391765 -0.04287046  x & y
#>  6:  1.36860228  1.71506499  2.05008469 -0.06191171  1.36860228  x & y
#>  7: -0.22577099  0.46091621 -0.49103117 -0.30596266 -0.22577099  x & y
#>  8:  1.51647060 -1.26506123 -2.30916888 -0.38047100  1.51647060  x & y
#>  9: -1.54875280 -0.68685285  1.00573852 -0.69470698 -1.54875280  x & y
#> 10:  0.58461375 -0.44566197 -0.70920076 -0.20791728  0.58461375  x & y
#> 11:  0.12385424  1.22408180 -0.68800862 -1.26539635  0.12385424  x & y
#> 12:  0.21594157  0.35981383  1.02557137  2.16895597  0.21594157  x & y
#> 13:  0.37963948  0.40077145 -0.28477301  1.20796200  0.37963948  x & y
#> 14: -0.50232345  0.11068272 -1.22071771 -1.12310858 -0.50232345  x & y
#> 15: -0.33320738 -0.55584113  0.18130348 -0.40288484 -0.33320738  x & y
#> 16: -1.01857538  1.78691314 -0.13889136 -0.46665535 -1.01857538  x & y
#>     row_index
#>         <int>
#>  1:         1
#>  2:         2
#>  3:         3
#>  4:         4
#>  5:         5
#>  6:         6
#>  7:         7
#>  8:         8
#>  9:         9
#> 10:        10
#> 11:        11
#> 12:        12
#> 13:        13
#> 14:        14
#> 15:        15
#> 16:        16

# 3. Toggle tolerance:
extract_diff_table(survey_data, survey_data_2, by=c('country', 'year'),
                    tolerance = 1e-5)
#> [1] 16
#>                diff indexes country  year variable1.x variable2.x variable3.x
#>              <char>  <char>  <char> <int>       <num>       <num>       <num>
#>  1: change_in_value       1       A  2010 -0.56047565   0.4978505  0.89512566
#>  2: change_in_value       2       A  2011 -0.23017749  -1.9666172  0.87813349
#>  3: change_in_value       3       A  2012  1.55870831   0.7013559  0.82158108
#>  4: change_in_value       4       A  2013  0.07050839  -0.4727914  0.68864025
#>  5: change_in_value       5       A  2014  0.12928774  -1.0678237  0.55391765
#>  6: change_in_value       6       A  2015  1.71506499  -0.2179749 -0.06191171
#>  7: change_in_value       7       A  2016  0.46091621  -1.0260044 -0.30596266
#>  8: change_in_value       8       A  2017 -1.26506123  -0.7288912 -0.38047100
#>  9: change_in_value       9       B  2010 -0.68685285  -0.6250393 -0.69470698
#> 10: change_in_value      10       B  2011 -0.44566197  -1.6866933 -0.20791728
#> 11: change_in_value      11       B  2012  1.22408180   0.8377870 -1.26539635
#> 12: change_in_value      12       B  2013  0.35981383   0.1533731  2.16895597
#> 13: change_in_value      13       B  2014  0.40077145  -1.1381369  1.20796200
#> 14: change_in_value      14       B  2015  0.11068272   1.2538149 -1.12310858
#> 15: change_in_value      15       B  2016 -0.55584113   0.4264642 -0.40288484
#> 16: change_in_value      16       B  2017  1.78691314  -0.2950715 -0.46665535
#>     variable4.x variable1.y variable2.y variable3.y variable4.y  .joyn
#>           <num>       <num>       <num>       <num>       <num> <fctr>
#>  1:  0.77996512 -0.56047565 -1.07179123  0.89512566  0.77996512  x & y
#>  2: -0.08336907 -0.23017749  0.30352864  0.87813349 -0.08336907  x & y
#>  3:  0.25331851  1.55870831  0.44820978  0.82158108  0.25331851  x & y
#>  4: -0.02854676  0.07050839  0.05300423  0.68864025 -0.02854676  x & y
#>  5: -0.04287046  0.12928774  0.92226747  0.55391765 -0.04287046  x & y
#>  6:  1.36860228  1.71506499  2.05008469 -0.06191171  1.36860228  x & y
#>  7: -0.22577099  0.46091621 -0.49103117 -0.30596266 -0.22577099  x & y
#>  8:  1.51647060 -1.26506123 -2.30916888 -0.38047100  1.51647060  x & y
#>  9: -1.54875280 -0.68685285  1.00573852 -0.69470698 -1.54875280  x & y
#> 10:  0.58461375 -0.44566197 -0.70920076 -0.20791728  0.58461375  x & y
#> 11:  0.12385424  1.22408180 -0.68800862 -1.26539635  0.12385424  x & y
#> 12:  0.21594157  0.35981383  1.02557137  2.16895597  0.21594157  x & y
#> 13:  0.37963948  0.40077145 -0.28477301  1.20796200  0.37963948  x & y
#> 14: -0.50232345  0.11068272 -1.22071771 -1.12310858 -0.50232345  x & y
#> 15: -0.33320738 -0.55584113  0.18130348 -0.40288484 -0.33320738  x & y
#> 16: -1.01857538  1.78691314 -0.13889136 -0.46665535 -1.01857538  x & y
#>     row_index
#>         <int>
#>  1:         1
#>  2:         2
#>  3:         3
#>  4:         4
#>  5:         5
#>  6:         6
#>  7:         7
#>  8:         8
#>  9:         9
#> 10:        10
#> 11:        11
#> 12:        12
#> 13:        13
#> 14:        14
#> 15:        15
#> 16:        16