Extract Different Values - User-facing - Table format Function to extract rows with different values between two dataframes.

Extract Different Values - User-facing - Table format Function to extract rows with different values between two dataframes.

Usage

extract_diff_table(
  dfx = NULL,
  dfy = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  myrror_object = NULL,
  output = c("simple", "full", "silent"),
  tolerance = 1e-07
)

Arguments

dfx

data.frame object.

dfy

data.frame object.

by

character, key to be used for dfx and dfy.

by.x

character, key to be used for dfx.

by.y

character, key to be used for dfy.

myrror_object

myrror object.

output

character, one of "simple", "full", "silent".

tolerance

numeric, default to 1e-7.

Value

data.table object with all observations for which at least 1 value is different.

Examples

extract_diff_table(iris, iris_var1)
#>           diff indexes Sepal.Length.x Sepal.Width.x Petal.Length.x
#>         <char>  <char>          <num>         <num>          <num>
#> 1: value_to_na      67            5.6           3.0            4.5
#> 2: value_to_na      80            5.7           2.6            3.5
#> 3: value_to_na      96            5.7           3.0            4.2
#> 4: value_to_na     104            6.3           2.9            5.6
#> 5: value_to_na     125            6.7           3.3            5.7
#>    Petal.Width.x  Species.x Sepal.Length.y Sepal.Width.y Petal.Length.y
#>            <num>     <char>          <num>         <num>          <num>
#> 1:           1.5 versicolor             NA           3.0            4.5
#> 2:           1.0 versicolor             NA           2.6            3.5
#> 3:           1.2 versicolor             NA           3.0            4.2
#> 4:           1.8  virginica             NA           2.9            5.6
#> 5:           2.1  virginica             NA           3.3            5.7
#>    Petal.Width.y  Species.y Petal.Area  .joyn row_index
#>            <num>     <char>      <num> <fctr>     <int>
#> 1:           1.5 versicolor       6.75  x & y        67
#> 2:           1.0 versicolor       3.50  x & y        80
#> 3:           1.2 versicolor       5.04  x & y        96
#> 4:           1.8  virginica      10.08  x & y       104
#> 5:           2.1  virginica      11.97  x & y       125
extract_diff_table(survey_data, survey_data_2, by=c('country', 'year'))
#>                diff indexes country  year variable1.x variable2.x variable3.x
#>              <char>  <char>  <char> <int>       <num>       <num>       <num>
#>  1: change_in_value       1       A  2010  0.54366717 -0.36741872   0.8270134
#>  2: change_in_value       2       A  2011 -0.18746702  1.08281824  -0.6280390
#>  3: change_in_value       3       A  2012  0.94976910 -0.88608285   0.1758058
#>  4: change_in_value       4       A  2013 -1.10673964 -1.35789652  -0.4435902
#>  5: change_in_value       5       A  2014  0.36097445  0.49807310  -0.3729658
#>  6: change_in_value       6       A  2015  2.03254692 -0.44915960   0.1220124
#>  7: change_in_value       7       A  2016 -0.28904570 -1.35026961   0.9950452
#>  8: change_in_value       8       A  2017  0.67921600  0.10604084  -0.1998401
#>  9: change_in_value       9       B  2010 -0.04356271  2.44402606   0.4113771
#> 10: change_in_value      10       B  2011  0.31081511 -0.08331262  -0.7091216
#> 11: change_in_value      11       B  2012  0.37805838 -0.83435863  -0.9148476
#> 12: change_in_value      12       B  2013  0.55137031  0.14181480   0.1287092
#> 13: change_in_value      13       B  2014  1.77946101 -1.27564125   1.2594642
#> 14: change_in_value      14       B  2015  0.06146327 -0.06316779  -0.6132742
#> 15: change_in_value      15       B  2016 -0.06867818 -1.28396057  -0.4870781
#> 16: change_in_value      16       B  2017  0.23823288  0.73642976  -0.5390066
#>      variable4.x variable1.y variable2.y variable3.y  variable4.y  .joyn
#>            <num>       <num>       <num>       <num>        <num> <fctr>
#>  1: -0.092927361  0.54366717  0.48571921   0.8270134 -0.092927361  x & y
#>  2:  0.672555322 -0.18746702 -0.26755620  -0.6280390  0.672555322  x & y
#>  3:  0.104421976  0.94976910 -1.22263500   0.1758058  0.104421976  x & y
#>  4: -0.464067726 -1.10673964 -0.04244568  -0.4435902 -0.464067726  x & y
#>  5:  0.000880445  0.36097445 -1.44511812  -0.3729658  0.000880445  x & y
#>  6:  0.242133191  2.03254692 -0.87467620   0.1220124  0.242133191  x & y
#>  7:  1.197393527 -0.28904570 -0.63024083   0.9950452  1.197393527  x & y
#>  8:  0.273731489  0.67921600  1.05829679  -0.1998401  0.273731489  x & y
#>  9: -1.017990476 -0.04356271 -0.77469057   0.4113771 -1.017990476  x & y
#> 10: -1.959103614  0.31081511 -0.56094771  -0.7091216 -1.959103614  x & y
#> 11: -0.075142008  0.37805838 -0.62187200  -0.9148476 -0.075142008  x & y
#> 12: -0.411169777  0.55137031 -0.45073978   0.1287092 -0.411169777  x & y
#> 13: -0.061513547  1.77946101  0.02804128   1.2594642 -0.061513547  x & y
#> 14: -0.018961455  0.06146327  0.92758765  -0.6132742 -0.018961455  x & y
#> 15:  1.667808010 -0.06867818 -0.19687494  -0.4870781  1.667808010  x & y
#> 16: -0.784008485  0.23823288 -0.16630616  -0.5390066 -0.784008485  x & y
#>     row_index
#>         <int>
#>  1:         1
#>  2:         2
#>  3:         3
#>  4:         4
#>  5:         5
#>  6:         6
#>  7:         7
#>  8:         8
#>  9:         9
#> 10:        10
#> 11:        11
#> 12:        12
#> 13:        13
#> 14:        14
#> 15:        15
#> 16:        16