Extract Different Values - User-facing - Table format Function to extract rows with different values between two dataframes.

Extract Different Values - User-facing - Table format Function to extract rows with different values between two dataframes.

Usage

extract_diff_table(
  dfx = NULL,
  dfy = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  myrror_object = NULL,
  output = c("simple", "full", "silent"),
  tolerance = 1e-07
)

Arguments

dfx

data.frame object.

dfy

data.frame object.

by

character, key to be used for dfx and dfy.

by.x

character, key to be used for dfx.

by.y

character, key to be used for dfy.

myrror_object

myrror object.

output

character, one of "simple", "full", "silent".

tolerance

numeric, default to 1e-7.

Value

data.table object with all observations for which at least 1 value is different.

Examples

extract_diff_table(iris, iris_var1)
#> # A tibble: 5 × 16
#>   variable     diff        indexes Sepal.Length.x Sepal.Width.x Petal.Length.x
#> * <fct>        <chr>       <chr>            <dbl>         <dbl>          <dbl>
#> 1 Sepal.Length value_to_na 67                 5.6           3              4.5
#> 2 Sepal.Length value_to_na 80                 5.7           2.6            3.5
#> 3 Sepal.Length value_to_na 96                 5.7           3              4.2
#> 4 Sepal.Length value_to_na 104                6.3           2.9            5.6
#> 5 Sepal.Length value_to_na 125                6.7           3.3            5.7
#> # ℹ 10 more variables: Petal.Width.x <dbl>, Species.x <chr>,
#> #   Sepal.Length.y <dbl>, Sepal.Width.y <dbl>, Petal.Length.y <dbl>,
#> #   Petal.Width.y <dbl>, Species.y <chr>, Petal.Area <dbl>, .joyn <fct>,
#> #   row_index <int>
extract_diff_table(survey_data, survey_data_2, by=c('country', 'year'))
#> # A tibble: 16 × 15
#>    variable  diff      indexes country  year variable1.x variable2.x variable3.x
#>  * <fct>     <chr>     <chr>   <chr>   <int>       <dbl>       <dbl>       <dbl>
#>  1 variable2 change_i… 1       A        2010      0.544      -0.367        0.827
#>  2 variable2 change_i… 2       A        2011     -0.187       1.08        -0.628
#>  3 variable2 change_i… 3       A        2012      0.950      -0.886        0.176
#>  4 variable2 change_i… 4       A        2013     -1.11       -1.36        -0.444
#>  5 variable2 change_i… 5       A        2014      0.361       0.498       -0.373
#>  6 variable2 change_i… 6       A        2015      2.03       -0.449        0.122
#>  7 variable2 change_i… 7       A        2016     -0.289      -1.35         0.995
#>  8 variable2 change_i… 8       A        2017      0.679       0.106       -0.200
#>  9 variable2 change_i… 9       B        2010     -0.0436      2.44         0.411
#> 10 variable2 change_i… 10      B        2011      0.311      -0.0833      -0.709
#> 11 variable2 change_i… 11      B        2012      0.378      -0.834       -0.915
#> 12 variable2 change_i… 12      B        2013      0.551       0.142        0.129
#> 13 variable2 change_i… 13      B        2014      1.78       -1.28         1.26 
#> 14 variable2 change_i… 14      B        2015      0.0615     -0.0632      -0.613
#> 15 variable2 change_i… 15      B        2016     -0.0687     -1.28        -0.487
#> 16 variable2 change_i… 16      B        2017      0.238       0.736       -0.539
#> # ℹ 7 more variables: variable4.x <dbl>, variable1.y <dbl>, variable2.y <dbl>,
#> #   variable3.y <dbl>, variable4.y <dbl>, .joyn <fct>, row_index <int>