Extract Different Values - User-facing - List format Function to extract rows with different values between two dataframes.

Extract Different Values - User-facing - List format Function to extract rows with different values between two dataframes.

Usage

extract_diff_values(
  dfx = NULL,
  dfy = NULL,
  by = NULL,
  by.x = NULL,
  by.y = NULL,
  myrror_object = NULL,
  output = c("simple", "full", "silent"),
  tolerance = 1e-07
)

Arguments

dfx

data.frame object.

dfy

data.frame object.

by

character, key to be used for dfx and dfy.

by.x

character, key to be used for dfx.

by.y

character, key to be used for dfy.

myrror_object

myrror object.

output

character, one of "simple", "full", "silent".

tolerance

numeric, default to 1e-7.

Value

list object with two items: diff_list and diff_table

Examples

extract_diff_values(iris, iris_var1)
#> $Sepal.Length
#> # A tibble: 5 × 4
#>   diff        indexes Sepal.Length.x Sepal.Length.y
#>   <chr>       <chr>            <dbl>          <dbl>
#> 1 value_to_na 104                6.3             NA
#> 2 value_to_na 125                6.7             NA
#> 3 value_to_na 67                 5.6             NA
#> 4 value_to_na 80                 5.7             NA
#> 5 value_to_na 96                 5.7             NA
#> 
extract_diff_values(survey_data, survey_data_2, by=c('country', 'year'))
#> $variable2
#> # A tibble: 16 × 4
#>    diff            indexes variable2.x variable2.y
#>    <chr>           <chr>         <dbl>       <dbl>
#>  1 change_in_value 1           -0.367       0.486 
#>  2 change_in_value 10          -0.0833     -0.561 
#>  3 change_in_value 11          -0.834      -0.622 
#>  4 change_in_value 12           0.142      -0.451 
#>  5 change_in_value 13          -1.28        0.0280
#>  6 change_in_value 14          -0.0632      0.928 
#>  7 change_in_value 15          -1.28       -0.197 
#>  8 change_in_value 16           0.736      -0.166 
#>  9 change_in_value 2            1.08       -0.268 
#> 10 change_in_value 3           -0.886      -1.22  
#> 11 change_in_value 4           -1.36       -0.0424
#> 12 change_in_value 5            0.498      -1.45  
#> 13 change_in_value 6           -0.449      -0.875 
#> 14 change_in_value 7           -1.35       -0.630 
#> 15 change_in_value 8            0.106       1.06  
#> 16 change_in_value 9            2.44       -0.775 
#>